+++ 楽しいエレクトロ・ジングル.score
@@ -1,0 +1,12 @@
+7 1 0 0 0 0 0 0 0 0 
+423 1 0 0 0 0 0 0 0 0 
+727 0 1 0 0 0 0 0 0 0 
+1054 0 0 1 0 0 0 0 0 0 
+1472 0 0 0 0 1 0 0 0 0 
+1889 0 0 0 0 0 1 0 0 0 
+2319 0 0 0 0 0 0 0 1 0 
+2743 0 0 0 0 0 0 1 0 0 
+3671 0 0 0 0 0 1 0 0 0 
+4534 0 0 0 1 0 0 0 0 0 
+4887 0 0 0 0 1 0 0 0 0 
+5302 0 0 1 0 0 0 0 0 0
